all rate PIDs working and motor output calculated ok

+++ Quadcopter/Receiver.h
@@ -1,18 +1,18 @@
 
 
-uint8_t addresses[][6] = {"1Node", "2Node"};
+byte addresses[][6] = {"1Node", "2Node"};
 bool radioNumber = 1; // receiver should be 1 (or anything not 0)
 RF24 radio(8, 7); // CE, CSN (SPI SS) *** WILL NEED TO BE UPDATED***
 
-uint8_t statusForAcknowledgement; // send this back to transmitter as acknowledgement package
+byte statusForAcknowledgement; // send this back to transmitter as acknowledgement package
 
 struct dataStruct {
-  uint16_t throttle; // number 1000 to 2000
-  uint16_t pitch;    // number 1000 to 2000
-  uint16_t roll;     // number 1000 to 2000
-  uint16_t yaw;     // number 1000 to 2000
-  uint8_t control; // for some control bits
-  uint8_t checksum;
+  int throttle; // number 1000 to 2000
+  int pitch;    // number 1000 to 2000
+  int roll;     // number 1000 to 2000
+  int yaw;     // number 1000 to 2000
+  byte control; // for some control bits
+  byte checksum;
 };
 
 struct dataStruct rcPackage;
